Fog is now created at startup.
Fog now isn't deleted when a room is left and re-created when it is entered again. The renderjob is instead removed, then re-added when the room is again entered.

Undid the changes from the last commit because they won't be useful once UVs for wall quads of fog are set properly.